Removed use of unit system (has semi-major flaw that I'll fix post-season), using literal conversion factors instead; fixed reset functionality on WrappingDouble; cleaned up code in Drive; added Shuffleboard accessors for PID parameters in Drive; added a bunch of temporary code in Robot and elsewhere.

+++ src/main/java/frc/team7170/robot2019/TeleopStateMachine.java
@@ -81,7 +81,7 @@ public class TeleopStateMachine {
     }
 
     private void drive(Event event) {
-        event.assertArgumentsTypes(Integer.class, Integer.class, Boolean.class);
+        event.assertArgumentsTypes(Double.class, Double.class, Boolean.class);
 
         double val0 = driveMultiplier * (double) event.arguments[0];
         double val1 = driveMultiplier * (double) event.arguments[1];
